pyproject.toml
[tool.poetry]
name = "bank-operations-coursework"
version = "0.1.0"
description = ""
authors = ["Artem Brazhnik <dambldor55@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
